Turn `polys.numberfields` from a module into a package.
As a part of this, we break this one large module into three parts:
* `minpoly`: everything to do with minimal polynomials
* `isomorphism`: number field isomorphisms
* `numbers`: isolating algebraic numbers, and rendering them in different forms
