+++ sympy/polys/numberfields/isomorphism.py
@@ -1,0 +1,151 @@
+"""Isomorphisms of number fields."""
+
+from sympy import S, AlgebraicNumber, Add, sympify
+from sympy.ntheory import sieve
+from sympy.polys.polytools import Poly, factor_list
+from sympy.utilities import public
+
+from mpmath import pslq, mp
+
+
+def is_isomorphism_possible(a, b):
+    """Returns `True` if there is a chance for isomorphism. """
+    n = a.minpoly.degree()
+    m = b.minpoly.degree()
+
+    if m % n != 0:
+        return False
+
+    if n == m:
+        return True
+
+    da = a.minpoly.discriminant()
+    db = b.minpoly.discriminant()
+
+    i, k, half = 1, m//n, db//2
+
+    while True:
+        p = sieve[i]
+        P = p**k
+
+        if P > half:
+            break
+
+        if ((da % p) % 2) and not (db % P):
+            return False
+
+        i += 1
+
+    return True
+
+
+def field_isomorphism_pslq(a, b):
+    """Construct field isomorphism using PSLQ algorithm. """
+    if not a.root.is_real or not b.root.is_real:
+        raise NotImplementedError("PSLQ doesn't support complex coefficients")
+
+    f = a.minpoly
+    g = b.minpoly.replace(f.gen)
+
+    n, m, prev = 100, b.minpoly.degree(), None
+
+    for i in range(1, 5):
+        A = a.root.evalf(n)
+        B = b.root.evalf(n)
+
+        basis = [1, B] + [ B**i for i in range(2, m) ] + [A]
+
+        dps, mp.dps = mp.dps, n
+        coeffs = pslq(basis, maxcoeff=int(1e10), maxsteps=1000)
+        mp.dps = dps
+
+        if coeffs is None:
+            break
+
+        if coeffs != prev:
+            prev = coeffs
+        else:
+            break
+
+        coeffs = [S(c)/coeffs[-1] for c in coeffs[:-1]]
+
+        while not coeffs[-1]:
+            coeffs.pop()
+
+        coeffs = list(reversed(coeffs))
+        h = Poly(coeffs, f.gen, domain='QQ')
+
+        if f.compose(h).rem(g).is_zero:
+            d, approx = len(coeffs) - 1, 0
+
+            for i, coeff in enumerate(coeffs):
+                approx += coeff*B**(d - i)
+
+            if A*approx < 0:
+                return [ -c for c in coeffs ]
+            else:
+                return coeffs
+        elif f.compose(-h).rem(g).is_zero:
+            return [ -c for c in coeffs ]
+        else:
+            n *= 2
+
+    return None
+
+
+def field_isomorphism_factor(a, b):
+    """Construct field isomorphism via factorization. """
+    _, factors = factor_list(a.minpoly, extension=b)
+
+    for f, _ in factors:
+        if f.degree() == 1:
+            coeffs = f.rep.TC().to_sympy_list()
+            d, terms = len(coeffs) - 1, []
+
+            for i, coeff in enumerate(coeffs):
+                terms.append(coeff*b.root**(d - i))
+
+            root = Add(*terms)
+
+            if (a.root - root).evalf(chop=True) == 0:
+                return coeffs
+
+            if (a.root + root).evalf(chop=True) == 0:
+                return [-c for c in coeffs]
+
+    return None
+
+
+@public
+def field_isomorphism(a, b, *, fast=True):
+    """Construct an isomorphism between two number fields. """
+    a, b = sympify(a), sympify(b)
+
+    if not a.is_AlgebraicNumber:
+        a = AlgebraicNumber(a)
+
+    if not b.is_AlgebraicNumber:
+        b = AlgebraicNumber(b)
+
+    if a == b:
+        return a.coeffs()
+
+    n = a.minpoly.degree()
+    m = b.minpoly.degree()
+
+    if n == 1:
+        return [a.root]
+
+    if m % n != 0:
+        return None
+
+    if fast:
+        try:
+            result = field_isomorphism_pslq(a, b)
+
+            if result is not None:
+                return result
+        except NotImplementedError:
+            pass
+
+    return field_isomorphism_factor(a, b)
