+++ sympy/polys/numberfields/minpoly.py
@@ -1,9 +1,9 @@
-"""Computational algebraic field theory. """
+"""Minimal polynomials for algebraic numbers."""
 
 from functools import reduce
 
 from sympy import (
-    S, Rational, AlgebraicNumber, GoldenRatio, TribonacciConstant,
+    S, Rational, GoldenRatio, TribonacciConstant,
     Add, Mul, sympify, Dummy, expand_mul, I, pi
 )
 from sympy.functions import sqrt, cbrt
@@ -12,7 +12,6 @@ from sympy.core.exprtools import Factors
 from sympy.core.function import _mexpand
 from sympy.functions.elementary.exponential import exp
 from sympy.functions.elementary.trigonometric import cos, sin, tan
-from sympy.ntheory import sieve
 from sympy.ntheory.factor_ import divisors
 from sympy.utilities.iterables import subsets
 
@@ -20,7 +19,6 @@ from sympy.polys.densetools import dup_eval
 from sympy.polys.domains import ZZ, QQ
 from sympy.polys.orthopolys import dup_chebyshevt
 from sympy.polys.polyerrors import (
-    IsomorphismFailed,
     NotAlgebraic,
     GeneratorsError,
 )
@@ -33,16 +31,11 @@ from sympy.polys.ring_series import rs_compose_add
 from sympy.polys.rings import ring
 from sympy.polys.rootoftools import CRootOf
 from sympy.polys.specialpolys import cyclotomic_poly
-from sympy.printing.lambdarepr import LambdaPrinter
-from sympy.printing.pycode import PythonCodePrinter, MpmathPrinter
 from sympy.simplify.radsimp import _split_gcd
 from sympy.simplify.simplify import _is_sum_surds
 from sympy.utilities import (
-    numbered_symbols, lambdify, public, sift
+    numbered_symbols, public, sift
 )
-
-from mpmath import pslq, mp
-
 
 
 def _choose_factor(factors, x, v, dom=QQ, prec=200, bound=5):
@@ -110,7 +103,7 @@ def _separate_sq(p):
 
     >>> from sympy import sqrt
     >>> from sympy.abc import x
-    >>> from sympy.polys.numberfields import _separate_sq
+    >>> from sympy.polys.numberfields.minpoly import _separate_sq
     >>> p= -x + sqrt(2) + sqrt(3) + sqrt(7)
     >>> p = _separate_sq(p); p
     -x**2 + 2*sqrt(3)*x + 2*sqrt(7)*x - 2*sqrt(21) - 8
@@ -173,7 +166,7 @@ def _minimal_polynomial_sq(p, n, x):
     Examples
     ========
 
-    >>> from sympy.polys.numberfields import _minimal_polynomial_sq
+    >>> from sympy.polys.numberfields.minpoly import _minimal_polynomial_sq
     >>> from sympy import sqrt
     >>> from sympy.abc import x
     >>> q = 1 + sqrt(2) + sqrt(3)
@@ -229,7 +222,7 @@ def _minpoly_op_algebraic_element(op, ex1, ex2, x, dom, mp1=None, mp2=None):
     ========
 
     >>> from sympy import sqrt, Add, Mul, QQ
-    >>> from sympy.polys.numberfields import _minpoly_op_algebraic_element
+    >>> from sympy.polys.numberfields.minpoly import _minpoly_op_algebraic_element
     >>> from sympy.abc import x, y
     >>> p1 = sqrt(sqrt(2) + 1)
     >>> p2 = sqrt(sqrt(2) - 1)
@@ -330,7 +323,7 @@ def _minpoly_pow(ex, pw, x, dom, mp=None):
     ========
 
     >>> from sympy import sqrt, QQ, Rational
-    >>> from sympy.polys.numberfields import _minpoly_pow, minpoly
+    >>> from sympy.polys.numberfields.minpoly import _minpoly_pow, minpoly
     >>> from sympy.abc import x, y
     >>> p = sqrt(1 + sqrt(2))
     >>> _minpoly_pow(p, 2, x, QQ)
@@ -918,229 +911,3 @@ def primitive_element(extension, x=None, *, ex=False, polys=False):
         return f.as_expr(), coeffs, H
     else:
         return f, coeffs, H
-
-
-def is_isomorphism_possible(a, b):
-    """Returns `True` if there is a chance for isomorphism. """
-    n = a.minpoly.degree()
-    m = b.minpoly.degree()
-
-    if m % n != 0:
-        return False
-
-    if n == m:
-        return True
-
-    da = a.minpoly.discriminant()
-    db = b.minpoly.discriminant()
-
-    i, k, half = 1, m//n, db//2
-
-    while True:
-        p = sieve[i]
-        P = p**k
-
-        if P > half:
-            break
-
-        if ((da % p) % 2) and not (db % P):
-            return False
-
-        i += 1
-
-    return True
-
-
-def field_isomorphism_pslq(a, b):
-    """Construct field isomorphism using PSLQ algorithm. """
-    if not a.root.is_real or not b.root.is_real:
-        raise NotImplementedError("PSLQ doesn't support complex coefficients")
-
-    f = a.minpoly
-    g = b.minpoly.replace(f.gen)
-
-    n, m, prev = 100, b.minpoly.degree(), None
-
-    for i in range(1, 5):
-        A = a.root.evalf(n)
-        B = b.root.evalf(n)
-
-        basis = [1, B] + [ B**i for i in range(2, m) ] + [A]
-
-        dps, mp.dps = mp.dps, n
-        coeffs = pslq(basis, maxcoeff=int(1e10), maxsteps=1000)
-        mp.dps = dps
-
-        if coeffs is None:
-            break
-
-        if coeffs != prev:
-            prev = coeffs
-        else:
-            break
-
-        coeffs = [S(c)/coeffs[-1] for c in coeffs[:-1]]
-
-        while not coeffs[-1]:
-            coeffs.pop()
-
-        coeffs = list(reversed(coeffs))
-        h = Poly(coeffs, f.gen, domain='QQ')
-
-        if f.compose(h).rem(g).is_zero:
-            d, approx = len(coeffs) - 1, 0
-
-            for i, coeff in enumerate(coeffs):
-                approx += coeff*B**(d - i)
-
-            if A*approx < 0:
-                return [ -c for c in coeffs ]
-            else:
-                return coeffs
-        elif f.compose(-h).rem(g).is_zero:
-            return [ -c for c in coeffs ]
-        else:
-            n *= 2
-
-    return None
-
-
-def field_isomorphism_factor(a, b):
-    """Construct field isomorphism via factorization. """
-    _, factors = factor_list(a.minpoly, extension=b)
-
-    for f, _ in factors:
-        if f.degree() == 1:
-            coeffs = f.rep.TC().to_sympy_list()
-            d, terms = len(coeffs) - 1, []
-
-            for i, coeff in enumerate(coeffs):
-                terms.append(coeff*b.root**(d - i))
-
-            root = Add(*terms)
-
-            if (a.root - root).evalf(chop=True) == 0:
-                return coeffs
-
-            if (a.root + root).evalf(chop=True) == 0:
-                return [-c for c in coeffs]
-
-    return None
-
-
-@public
-def field_isomorphism(a, b, *, fast=True):
-    """Construct an isomorphism between two number fields. """
-    a, b = sympify(a), sympify(b)
-
-    if not a.is_AlgebraicNumber:
-        a = AlgebraicNumber(a)
-
-    if not b.is_AlgebraicNumber:
-        b = AlgebraicNumber(b)
-
-    if a == b:
-        return a.coeffs()
-
-    n = a.minpoly.degree()
-    m = b.minpoly.degree()
-
-    if n == 1:
-        return [a.root]
-
-    if m % n != 0:
-        return None
-
-    if fast:
-        try:
-            result = field_isomorphism_pslq(a, b)
-
-            if result is not None:
-                return result
-        except NotImplementedError:
-            pass
-
-    return field_isomorphism_factor(a, b)
-
-
-@public
-def to_number_field(extension, theta=None, *, gen=None):
-    """Express `extension` in the field generated by `theta`. """
-    if hasattr(extension, '__iter__'):
-        extension = list(extension)
-    else:
-        extension = [extension]
-
-    if len(extension) == 1 and isinstance(extension[0], tuple):
-        return AlgebraicNumber(extension[0])
-
-    minpoly, coeffs = primitive_element(extension, gen, polys=True)
-    root = sum([ coeff*ext for coeff, ext in zip(coeffs, extension) ])
-
-    if theta is None:
-        return AlgebraicNumber((minpoly, root))
-    else:
-        theta = sympify(theta)
-
-        if not theta.is_AlgebraicNumber:
-            theta = AlgebraicNumber(theta, gen=gen)
-
-        coeffs = field_isomorphism(root, theta)
-
-        if coeffs is not None:
-            return AlgebraicNumber(theta, coeffs)
-        else:
-            raise IsomorphismFailed(
-                "%s is not in a subfield of %s" % (root, theta.root))
-
-
-class IntervalPrinter(MpmathPrinter, LambdaPrinter):
-    """Use ``lambda`` printer but print numbers as ``mpi`` intervals. """
-
-    def _print_Integer(self, expr):
-        return "mpi('%s')" % super(PythonCodePrinter, self)._print_Integer(expr)
-
-    def _print_Rational(self, expr):
-        return "mpi('%s')" % super(PythonCodePrinter, self)._print_Rational(expr)
-
-    def _print_Half(self, expr):
-        return "mpi('%s')" % super(PythonCodePrinter, self)._print_Rational(expr)
-
-    def _print_Pow(self, expr):
-        return super(MpmathPrinter, self)._print_Pow(expr, rational=True)
-
-@public
-def isolate(alg, eps=None, fast=False):
-    """Give a rational isolating interval for an algebraic number. """
-    alg = sympify(alg)
-
-    if alg.is_Rational:
-        return (alg, alg)
-    elif not alg.is_real:
-        raise NotImplementedError(
-            "complex algebraic numbers are not supported")
-
-    func = lambdify((), alg, modules="mpmath", printer=IntervalPrinter())
-
-    poly = minpoly(alg, polys=True)
-    intervals = poly.intervals(sqf=True)
-
-    dps, done = mp.dps, False
-
-    try:
-        while not done:
-            alg = func()
-
-            for a, b in intervals:
-                if a <= alg.a and alg.b <= b:
-                    done = True
-                    break
-            else:
-                mp.dps *= 2
-    finally:
-        mp.dps = dps
-
-    if eps is not None:
-        a, b = poly.refine_root(a, b, eps=eps, fast=fast)
-
-    return (a, b)
